moved handler code to a saperate file
