+++ services/ip_cam_trendnet_tv_ip100/consts.py
@@ -1,0 +1,10 @@
+DEFAULT_PORT = 80
+EVENT_TYPE_FIELD_NAME = "event_type"
+TRENDNET_ADMIN_ACCESS_EVENT = "trendnet_tv_ip100_admin_access"
+TRENDNET_ADMIN_POST_ATTEMPT = "trendnet_tv_ip100_post_attempt"
+ORIGINATING_IP_FIELD_NAME = "originating_ip"
+ORIGINATING_PORT_FIELD_NAME = "originating_port"
+REQUEST_FIELD_NAME = "request"
+DEFAULT_SERVER_VERSION = "Camera Web Server/1.0"
+CAMERA_IMAGE_PATH = "/image.jpg"
+DEFAULT_CONTENT_TYPE = "image/jpeg"
